Now changing levels on dragon log adaptor does not change all levels
